Refactoring to separate cloud and local RPC style requests
The session considers the requests as procedure calls that return a result, but the implementation between cloud and local is very different.  Local requests do return the result inline in the HTTP response, but cloud requests return an empty JSON response and return the response from the scanner asynchronously through MQTT.

This isn't working yet - the scanner reports an invalid privet token - but should be completely implemented for local scanners.